adding all seeds functions

+++ api/src/database/seeders/seed.ts
@@ -1,29 +1,50 @@
 import { config } from '../../config/env'
 import { db } from '..'
 
-import { createUsers } from './users'
+import { createCountries } from './countries'
 import { createTags } from './tags'
+import { createUsers } from './users'
 import { createAwards } from './awards'
 import { createBrands } from './brands'
 import { createProducts } from './products'
 
 async function main(): Promise<void> {
-    if (config.NODE_ENV === 'development') {
+    const countriesCount = await db.country.count()
+    const usersCount = await db.user.count()
+    const tagsCount = await db.tag.count()
+    const awardsCount = await db.award.count()
+    const brandsCount = await db.brand.count()
+    const productsCount = await db.product.count()
+
+    if (countriesCount !== 250) {
+        await db.country.deleteMany()
+        await createCountries()
+    }
+
+    if (usersCount < 17 && config.NODE_ENV === 'development') {
         await db.user.deleteMany()
         await createUsers()
     }
 
-    await db.award.deleteMany()
-    await db.brand.deleteMany()
-    await db.tag.deleteMany()
-    await db.product.deleteMany()
+    if (tagsCount !== 19) {
+        await db.tag.deleteMany()
+        await createTags()
+    }
 
-    await createAwards()
-    await createBrands()
-    await createTags()
-    await createProducts()
+    if (awardsCount !== 9) {
+        await db.award.deleteMany()
+        await createAwards()
+    }
 
-    console.info('Seeding completed 🌱')
+    if (brandsCount !== 29) {
+        await db.brand.deleteMany()
+        await createBrands()
+    }
+
+    if (productsCount !== 49) {
+        await db.product.deleteMany()
+        await createProducts()
+    }
 }
 
 try {
